🎨 improve navigation menu with context management and button component integration
